finish

+++ main.cpp
@@ -1,12 +1,12 @@
 #include<iostream>
 #include<locale.h>
 #include"interface.h"
-#include <time.h>
+#include<ctime>
 
 using namespace std;
 
 int main(void) {
-	srand(time(NULL));
+	srand(static_cast<unsigned int>(time(NULL)));
 	setlocale(LC_ALL, "RU");
 
 	Greatings();
